fixed formatting issue with interactive message
