+++ src/containers/InteractiveMessage/InteractiveMessage.tsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect, useMemo } from 'react';
 import * as Yup from 'yup';
 import { useTranslation } from 'react-i18next';
 import { EditorState } from 'draft-js';
-import axios from 'axios';
 import { useLazyQuery, useQuery } from '@apollo/client';
 
 import { ReactComponent as InteractiveMessageIcon } from 'assets/images/icons/InteractiveMessage/Dark.svg';
@@ -22,11 +21,16 @@ import { LanguageBar } from 'components/UI/LanguageBar/LanguageBar';
 import { LIST, MEDIA_MESSAGE_TYPES, QUICK_REPLY } from 'common/constants';
 import { validateMedia } from 'common/utils';
 import Loading from 'components/UI/Layout/Loading/Loading';
-import { WhatsAppToDraftEditor } from 'common/RichEditor';
-import { FLOW_EDITOR_API } from 'config';
-import { getAuthSession } from 'services/AuthService';
+import { convertToWhatsApp, WhatsAppToDraftEditor } from 'common/RichEditor';
 import { InteractiveOptions } from './InteractiveOptions/InteractiveOptions';
 import styles from './InteractiveMessage.module.css';
+import {
+  convertJSONtoStateData,
+  getDefaultValuesByTemplate,
+  getPayloadByMediaType,
+  getVariableOptions,
+  validator,
+} from './InteractiveMessage.helper';
 
 export interface FlowProps {
   match: any;
@@ -39,96 +43,6 @@ const queries = {
   createItemQuery: CREATE_INTERACTIVE,
   updateItemQuery: UPDATE_INTERACTIVE,
   deleteItemQuery: DELETE_INTERACTIVE,
-};
-
-const convertJSONtoStateData = (JSONData: any, interactiveType: string) => {
-  const data = { ...JSONData };
-  const { title, body, items, content, options, globalButtons } = data;
-
-  if (interactiveType === QUICK_REPLY) {
-    const { type, header, url, text } = content;
-    const result: any = {};
-    result.templateButtons = options.map((option: any) => ({ value: option.title }));
-    result.title = header || '';
-    switch (type) {
-      case 'image':
-      case 'video':
-        result.type = `${type.toUpperCase()}`;
-        result.attachmentURL = url;
-        result.body = text;
-        break;
-      case 'file':
-        result.type = 'DOCUMENT';
-        result.attachmentURL = url;
-        break;
-      default:
-        result.type = null;
-        result.body = text || '';
-    }
-    return result;
-  }
-
-  const result: any = {};
-  result.templateButtons = items.map((item: any) => {
-    const itemOptions = item.options.map((option: any) => ({
-      title: option.title,
-      description: option.description,
-    }));
-    return {
-      title: item.title,
-      options: itemOptions,
-    };
-  });
-  result.body = body;
-  result.title = title;
-  result.globalButton = globalButtons[0].title;
-  return result;
-};
-
-const getDefaultValuesByTemplate = (templateData: any) => {
-  const { type: templateType, interactiveContent } = templateData;
-  const data = JSON.parse(interactiveContent);
-
-  let result: any = {};
-  if (templateType === QUICK_REPLY) {
-    const { type, content, options } = data;
-    const updatedOptions = options.map(() => ({ type: 'text', title: '' }));
-    const updatedContent = Object.keys(content).reduce((res: any, key: string) => {
-      if (['type', 'url'].includes(key)) res[key] = content[key];
-      else res[key] = '';
-      return res;
-    }, {});
-
-    result.type = type;
-    result.content = updatedContent;
-    result.options = updatedOptions;
-  }
-
-  if (templateType === LIST) {
-    result = Object.keys(data).reduce((res: any, key: string) => {
-      const dataVal = data[key];
-      if (typeof dataVal === 'string') {
-        res[key] = '';
-      }
-
-      if (key === 'items') {
-        const items: any = dataVal.map((item: any) => {
-          const optionVal = { type: 'text', title: '', description: '' };
-          const { options } = item;
-          const updatedOptions = options.map(() => optionVal);
-          return { title: '', subtitle: '', options: updatedOptions };
-        });
-        res[key] = items;
-      }
-
-      if (key === 'globalButtons') {
-        res[key] = dataVal.map(() => ({ type: 'text', title: '' }));
-      }
-
-      return res;
-    }, {});
-  }
-  return result;
 };
 
 export const InteractiveMessage: React.SFC<FlowProps> = ({ match }) => {
@@ -150,49 +64,16 @@ export const InteractiveMessage: React.SFC<FlowProps> = ({ match }) => {
   const [previousState, setPreviousState] = useState<any>({});
   const [warning, setWarning] = useState<any>();
 
-  useEffect(() => {
-    const glificBase = FLOW_EDITOR_API;
-    const contactFieldsprefix = '@contact.fields.';
-    const contactVariablesprefix = '@contact.';
-    const headers = { Authorization: getAuthSession('access_token') };
-
-    const getVariableOptions = async () => {
-      // get fields keys
-      const fieldsData = await axios.get(`${glificBase}fields`, {
-        headers,
-      });
-
-      const fields = fieldsData.data.results.map((i: any) => contactFieldsprefix.concat(i.key));
-
-      // get contact keys
-      const contactData = await axios.get(`${glificBase}completion`, {
-        headers,
-      });
-
-      const properties = contactData.data.context.types.find(
-        ({ name }: { name: string }) => name === 'contact'
-      );
-
-      const contacts =
-        properties &&
-        properties.properties
-          .map((i: any) => contactVariablesprefix.concat(i.key))
-          .concat(fields)
-          .map((val: string) => ({ name: val }))
-          .slice(1);
-
-      setContactVariables(contacts);
-    };
-
-    getVariableOptions();
-  }, []);
-
   const { data: languages } = useQuery(USER_LANGUAGES, {
     variables: { opts: { order: 'ASC' } },
   });
 
   const [getInteractiveTemplateById, { data: template }] =
     useLazyQuery<any>(GET_INTERACTIVE_MESSAGE);
+
+  useEffect(() => {
+    getVariableOptions(setContactVariables);
+  }, []);
 
   useEffect(() => {
     if (languages) {
@@ -500,31 +381,6 @@ export const InteractiveMessage: React.SFC<FlowProps> = ({ match }) => {
     },
   ];
 
-  const getPayloadByMediaType = (mediaType: string, payload: any) => {
-    const result: any = {};
-
-    switch (mediaType) {
-      case 'IMAGE':
-      case 'VIDEO':
-        result.type = `${mediaType.toLowerCase()}`;
-        result.url = payload.attachmentURL;
-        result.text = payload.body.getCurrentContent().getPlainText();
-        break;
-      case 'DOCUMENT':
-        result.type = 'file';
-        result.url = payload.attachmentURL;
-        result.filename = 'file';
-        break;
-      default:
-        result.type = 'text';
-        result.header = payload.title;
-        result.text = payload.body.getCurrentContent().getPlainText();
-        break;
-    }
-
-    return result;
-  };
-
   const getTemplateButtonPayload = (typeVal: string, buttons: Array<any>) => {
     if (typeVal === QUICK_REPLY) {
       return buttons.map((button: any) => ({ type: 'text', title: button.value }));
@@ -574,7 +430,7 @@ export const InteractiveMessage: React.SFC<FlowProps> = ({ match }) => {
     }
 
     if (templateTypeVal === LIST) {
-      const bodyText = payload.body.getCurrentContent().getPlainText();
+      const bodyText = convertToWhatsApp(payload.body);
       const items = getTemplateButtonPayload(templateTypeVal, templateButtonVal);
       const globalButtons = [{ type: 'text', title: globalButtonVal }];
 
@@ -686,64 +542,7 @@ export const InteractiveMessage: React.SFC<FlowProps> = ({ match }) => {
 
   const formFields = templateType === LIST ? [...fields] : [...fields, ...attachmentInputs];
 
-  const validation: any = {
-    title: Yup.string()
-      .required(t('Title is required'))
-      .max(60, t('Title can be at most 60 characters')),
-    body: Yup.string()
-      .transform((current, original) => original.getCurrentContent().getPlainText())
-      .required(t('Message content is required.')),
-  };
-
-  if (templateType === LIST) {
-    validation.templateButtons = Yup.array()
-      .of(
-        Yup.object().shape({
-          title: Yup.string()
-            .required(t('Required'))
-            .max(24, t('Section title can be at most 24 characters')),
-          options: Yup.array().of(
-            Yup.object().shape({
-              title: Yup.string()
-                .required(t('Title is required'))
-                .max(24, t('Title can be at most 24 characters')),
-              description: Yup.string().max(72, t('Description can be at most 72 characters')),
-            })
-          ),
-        })
-      )
-      .min(1);
-
-    validation.globalButton = Yup.string()
-      .required(t('Required'))
-      .max(20, t('Button value can be at most 20 characters'));
-  } else {
-    validation.templateButtons = Yup.array()
-      .of(
-        Yup.object().shape({
-          value: Yup.string()
-            .required(t('Required'))
-            .max(20, t('Button value can be at most 20 characters')),
-        })
-      )
-      .min(1)
-      .max(3);
-
-    validation.type = Yup.object()
-      .nullable()
-      .when('attachmentURL', {
-        is: (val: string) => val && val !== '',
-        then: Yup.object().nullable().required(t('Type is required.')),
-      });
-
-    validation.attachmentURL = Yup.string()
-      .nullable()
-      .when('type', {
-        is: (val: any) => val && val.id,
-        then: Yup.string().required(t('Attachment URL is required.')),
-      });
-  }
-
+  const validation = validator(templateType, t);
   const validationScheme = Yup.object().shape(validation, [['type', 'attachmentURL']]);
 
   const getPreviewData = () => {
